выделен класс SessionHelper для управления сессиями
