+++ fixture/application.py
@@ -1,4 +1,5 @@
 from selenium.webdriver.chrome.webdriver import WebDriver
+from fixture.session import SessionHelper
 
 
 class Application:
@@ -6,21 +7,11 @@ class Application:
     def __init__(self):
         self.wd = WebDriver()
         self.wd.implicitly_wait(30)
+        self.session = SessionHelper(self)
 
     def open_home_page(self):
         wd = self.wd
         wd.get("http://localhost/addressbook/")
-
-    def login(self, user_name, password):
-        wd = self.wd
-        self.open_home_page()
-        wd.find_element_by_name("user").click()
-        wd.find_element_by_name("user").clear()
-        wd.find_element_by_name("user").send_keys(user_name)
-        wd.find_element_by_name("pass").click()
-        wd.find_element_by_name("pass").clear()
-        wd.find_element_by_name("pass").send_keys(password)
-        wd.find_element_by_id("LoginForm").submit()
 
     def open_group_page(self):
         wd = self.wd
@@ -48,10 +39,6 @@ class Application:
     def return_to_group_page(self):
         wd = self.wd
         wd.find_element_by_link_text("group page").click()
-
-    def logout(self):
-        wd = self.wd
-        wd.find_element_by_link_text("Logout").click()
 
     def destroy(self):
         self.wd.quit()
